perf(block): rewire the engine onto the journal — carve dispatcher, cold-read hydrate, self-heal
Replace the chunk-at-ingest rollup + pending-carve machinery (carver.go)
with a background dispatcher that drives journal.Carve through the
engineDeduper/engineBlockSink carve targets; the BlockSink writes the
per-file FileChunk manifest rows inside the DefaultCommitBlock txn.

- Carve dispatcher (carve_dispatch.go) calls local.Carve on a ticker,
  gated by carveActive; SetCarveTargets wired once all deps are present.
- Writes go straight to journal.WriteAt; Flush -> journal.Commit;
  DrainRollups -> Carve(Force); ResetLocalState -> ListFiles+Delete.
- Reads resolve warm via the journal interval index; cold reads fetch
  covering blocks from the remote, verify BLAKE3, and Hydrate by
  (file,offset), then re-read. Self-heal redesigned per (file,offset).
- GC/refcount reap stays on gc_block.go; the journal's own GC is never
  driven from the carve loop. Delete the hash-keyed local GC (LocalStore
  returns nil so the local-tier sweep no-ops).
- Retire tests for the deleted machinery; keep durability #1274 and the
  carve-seam acceptance test; BlockStoreConformance runs through the real
  Syncer with a journal-backed FSStore.

+++ pkg/block/engine/dataextents.go
@@ -27,8 +27,8 @@ func (bs *Store) DataExtents(ctx context.Context, payloadID string, fileSize uin
 		return nil, nil
 	}
 
-	// (a) Pre-rollup bytes the local tier knows but CAS does not yet.
-	ext, err := bs.local.DataExtents(ctx, payloadID, fileSize)
+	// (a) Bytes the local journal tier knows (dirty or evicted/cold).
+	ext, err := bs.local.DataExtents(ctx, payloadID, int64(fileSize))
 	if err != nil {
 		return nil, err
 	}
